Backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.4
psycopg2-binary==2.9.9
